Replaced float with double

+++ src/Example.java
@@ -6,8 +6,8 @@ public class Example {
 
     public static void main(String[] args) throws Exception {
         System.out.print("Please enter a number you want to find the absolute value of: ");
-        float value = scanner.nextFloat();
-        float absoluteValue = findAbsoluteValue.findValue(value);
+        double value = scanner.nextFloat();
+        double absoluteValue = findAbsoluteValue.findValue(value);
         System.out.println("|" + value + "| = " + absoluteValue);
         System.out.println("The absolute value of " + value + " is " + absoluteValue);
     }
